+++ src/client/popup-window/configs.js
@@ -30,6 +30,17 @@ export const windowDefaultMinStyle = {
   top: "unset",
   left: "unset"
 };
+export const windowTitleDefaultStyle = {
+  height: "20px",
+  textAlign: "center"
+};
+export const windowBodyDefaultStyle = {
+  height: "100%",
+  margin: "10px",
+  display: "flex",
+  border: "0.5px solid rgba(0, 0, 0, 0.1)",
+  flexFlow: "row wrap"
+};
 export const closeBtnStyle = {
   color: "#E6605C",
   fontSize: "14px"